+++ R/crew_package.R
@@ -19,6 +19,7 @@
 #'   [`clustermq`](https://mschubert.github.io/clustermq/),
 #'   and [`batchtools`](https://github.com/mlr-org/batchtools).
 #' @importFrom cli cli_progress_bar cli_progress_done cli_progress_update
+#' @importFrom collections priority_queue
 #' @importFrom data.table rbindlist
 #' @importFrom later later run_now
 #' @importFrom mirai call_mirai daemon daemons is_mirai is_mirai_error